requirements.txt
# LLM Integration - Supports multiple providers
litellm>=1.17.0
openai>=1.0.0

# Azure Identity (DefaultAzureCredential)
azure-identity>=1.15.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# UI Components
gradio>=4.0.0
rich>=13.0.0
markdown>=3.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
